# docs/swagger.yaml
basePath: /
definitions:
  github_com_telkomdev-candrasaka_hris-monitoring_git_internal_repositories.AttendanceRiskReportRow:
    properties:
      approved_leave_count:
        type: integer
      attendance_records:
        type: integer
      attendance_risk_score:
        type: number
      city:
        type: string
      estimated_available_staff:
        type: integer
      headcount:
        type: integer
      location_id:
        type: integer
      location_name:
        type: string
      location_type:
        type: string
      minimum_staffing:
        type: integer
      open_attendance_count:
        type: integer
      pending_leave_count:
        type: integer
      staffing_risk:
        type: string
    type: object
  github_com_telkomdev-candrasaka_hris-monitoring_git_internal_repositories.LabourCostLeakageReportRow:
    properties:
      city:
        type: string
      headcount:
        type: integer
      leakage_percentage:
        type: number
      location_id:
        type: integer
      location_name:
        type: string
      location_type:
        type: string
      total_absent_count:
        type: integer
      total_base_salary:
        type: number
      total_deduction:
        type: number
      total_gross_salary:
        type: number
      total_late_count:
        type: integer
      total_net_salary:
        type: number
    type: object
  internal_handlers.AssetResponseDoc:
    properties:
      category:
        example: jaket thermal
        type: string
      condition:
        example: Layak
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Jaket Thermal A1
        type: string
      notes:
        example: Untuk staff cold storage
        type: string
      serial_number:
        example: APD-001
        type: string
      status:
        example: borrowed
        type: string
      user_id:
        example: 10
        type: integer
    type: object
  internal_handlers.AttendanceResponseDoc:
    properties:
      check_in:
        example: "2026-05-23T08:05:00Z"
        type: string
      check_out:
        example: "2026-05-23T17:30:00Z"
        type: string
      device_latitude:
        example: -6.2
        type: number
      device_longitude:
        example: 106.9
        type: number
      id:
        example: 1
        type: integer
      location_id:
        example: 1
        type: integer
      selfie_path:
        example: uploads/selfies/user_10_123.jpg
        type: string
      status:
        example: present_checked_out
        type: string
      user_id:
        example: 10
        type: integer
    type: object
  internal_handlers.AttendanceRiskReportListDoc:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_telkomdev-candrasaka_hris-monitoring_git_internal_repositories.AttendanceRiskReportRow'
        type: array
    type: object
...
  internal_handlers.CreateAssetRequestDoc:
    properties:
      category:
        example: jaket thermal
        type: string
      condition:
        example: Layak
        type: string
      name:
        example: Jaket Thermal A1
        type: string
      notes:
        example: Untuk staff cold storage
        type: string
      serial_number:
        example: APD-001
        type: string
    type: object
  internal_handlers.CreateUserRequestDoc:
    properties:
      base_salary:
        example: 5000000
        type: number
      email:
        example: budi@example.com
        type: string
      location_id:
        example: 1
        type: integer
      name:
        example: Budi
        type: string
      password:
        example: secret123
        type: string
      role:
        example: staff_gudang
        type: string
      shift_id:
        example: 2
        type: integer
    type: object
  internal_handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  internal_handlers.LabourCostLeakageReportListDoc:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_telkomdev-candrasaka_hris-monitoring_git_internal_repositories.LabourCostLeakageReportRow'
        type: array
    type: object
...
  internal_handlers.LeaveResponseDoc:
    properties:
      approved_at:
        example: "2026-05-23T10:00:00Z"
        type: string
      approved_by:
        example: 1
        type: integer
      document_path:
        example: uploads/leaves/leave_10_123.pdf
        type: string
      end_date:
        example: "2026-06-03T00:00:00Z"
        type: string
      id:
        example: 5
        type: integer
      leave_type:
        example: sakit
        type: string
      reason:
        example: Demam tinggi
        type: string
      start_date:
        example: "2026-06-01T00:00:00Z"
        type: string
      status:
        example: approved
        type: string
      user_id:
        example: 10
        type: integer
    type: object
  internal_handlers.LocationPayloadDoc:
    properties:
      address:
        example: Jl. Raya Bekasi No. 1
        type: string
      city:
        example: Bekasi
        type: string
      geofence_radius:
        example: 100
        type: number
      latitude:
        example: -6.2
        type: number
      longitude:
        example: 106.9
        type: number
      minimum_staffing:
        example: 8
        type: integer
      name:
        example: Warehouse Bekasi
        type: string
      province:
        example: Jawa Barat
        type: string
      type:
        example: warehouse
        type: string
    type: object
  internal_handlers.LocationResponseDoc:
    properties:
      address:
        example: Jl. Raya Bekasi No. 1
        type: string
      city:
        example: Bekasi
        type: string
      geofence_radius:
        example: 100
        type: number
      id:
        example: 1
        type: integer
      latitude:
        example: -6.2
        type: number
      longitude:
        example: 106.9
        type: number
      minimum_staffing:
        example: 8
        type: integer
      name:
        example: Warehouse Bekasi
        type: string
      province:
        example: Jawa Barat
        type: string
      type:
        example: warehouse
        type: string
    type: object
...
  internal_handlers.PayrollResponseDoc:
    properties:
      absent_count:
        example: 0
        type: integer
      base_salary:
        example: 5000000
        type: number
      cold_storage_allowance:
        example: 500000
        type: number
      generated_at:
        example: "2026-05-23T10:00:00Z"
        type: string
      gross_salary:
        example: 5625000
        type: number
      id:
        example: 1
        type: integer
      late_count:
        example: 1
        type: integer
      month:
        example: 5
        type: integer
      net_salary:
        example: 5575000
        type: number
      overtime_pay:
        example: 125000
        type: number
      total_attendance:
        example: 22
        type: integer
      total_deduction:
        example: 50000
        type: number
      user_id:
        example: 10
        type: integer
      year:
        example: 2026
        type: integer
    type: object
...
  internal_handlers.ShiftResponseDoc:
    properties:
      cross_midnight:
        example: false
        type: boolean
      end_time:
        example: "17:00"
        type: string
      grace_minutes:
        example: 15
        type: integer
      id:
        example: 2
        type: integer
      location_id:
        example: 1
        type: integer
      name:
        example: Shift Pagi
        type: string
      start_time:
        example: "08:00"
        type: string
    type: object
...
  internal_handlers.UserResponseDoc:
    properties:
      base_salary:
        example: 5000000
        type: number
      email:
        example: budi@example.com
        type: string
      id:
        example: 10
        type: integer
      location:
        $ref: '#/definitions/internal_handlers.LocationResponseDoc'
      location_id:
        example: 1
        type: integer
      name:
        example: Budi
        type: string
      role:
        example: staff_gudang
        type: string
      shift:
        $ref: '#/definitions/internal_handlers.ShiftResponseDoc'
      shift_id:
        example: 2
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Backend HRIS dan operasional frozen-food retail dengan JWT auth, geofenced
    attendance + selfie, leave approval with staffing risk, payroll, PPE/asset compliance,
    dan executive reports.
  title: HRIS Monitoring API
  version: "1.0"
paths:
  /api/reports/attendance-risk:
    get:
      description: Executive report showing attendance anomalies, leave load, minimum
        staffing pressure, and computed attendance risk by location.
      parameters:
      - description: Month
        in: query
        name: month
        type: integer
      - description: Year
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.AttendanceRiskReportListDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get attendance risk report
      tags:
      - Reports
  /api/reports/labour-cost-leakage:
    get:
      description: Executive report showing attendance-related payroll leakage by
        location, including deductions, lateness, absence counts, and leakage percentage.
      parameters:
      - description: Month
        in: query
        name: month
        type: integer
      - description: Year
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.LabourCostLeakageReportListDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get labour cost leakage report
      tags:
      - Reports
  /assets:
    get:
      description: List operational assets and PPE items including status, condition,
        and borrower linkage.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.AssetResponseDoc'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List assets
      tags:
      - Assets
    post:
      consumes:
      - application/json
      description: Create a new PPE or operational asset record.
      parameters:
      - description: Create asset payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateAssetRequestDoc'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.AssetResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create asset
      tags:
      - Assets
...
  /attendance:
    get:
      description: Return attendance history for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.AttendanceResponseDoc'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List current user's attendances
      tags:
      - Attendance
...
  /attendance/checkin:
    post:
      consumes:
      - multipart/form-data
      description: Check in with location validation, geofence validation, selfie
        upload, shift-aware lateness, and warehouse PPE compliance.
      parameters:
      - description: Location ID
        in: formData
        name: location_id
        required: true
        type: integer
      - description: Device latitude
        in: formData
        name: latitude
        required: true
        type: number
      - description: Device longitude
        in: formData
        name: longitude
        required: true
        type: number
      - description: Selfie image
        in: formData
        name: selfie
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.AttendanceResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Check in attendance
      tags:
      - Attendance
  /attendance/checkout:
    post:
      description: Close the current open attendance record for the authenticated
        user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.AttendanceResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Check out attendance
      tags:
      - Attendance
  /leaves:
    get:
      description: Return leave history for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.LeaveResponseDoc'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List current user's leaves
      tags:
      - Leaves
    post:
      consumes:
      - multipart/form-data
      description: Submit leave or permit request with optional supporting document
        upload.
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: formData
        name: start_date
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: formData
        name: end_date
        required: true
        type: string
      - description: Leave type
        in: formData
        name: leave_type
        required: true
        type: string
      - description: Reason
        in: formData
        name: reason
        type: string
      - description: Supporting document
        in: formData
        name: document
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.LeaveResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Apply leave
      tags:
      - Leaves
...
  /locations:
    get:
      description: Get all registered locations including type, geofence, and minimum
        staffing.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.LocationResponseDoc'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List locations
      tags:
      - Locations
    post:
      consumes:
      - application/json
      description: Create a new outlet or warehouse location used for geofencing,
        staffing rules, and location-driven payroll logic.
      parameters:
      - description: Location payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.LocationPayloadDoc'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.LocationResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create location
      tags:
      - Locations
...
  /payrolls:
    get:
      description: Calculate current user's payroll for a given month and year, including
        warehouse allowance and overtime.
      parameters:
      - description: Month
        in: query
        name: month
        type: integer
      - description: Year
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.PayrollResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Calculate or get payroll
      tags:
      - Payroll
  /payrolls/download:
    get:
      description: Generate and download a PDF payslip for the authenticated user
        and selected month/year.
      parameters:
      - description: Month
        in: query
        name: month
        type: integer
      - description: Year
        in: query
        name: year
        type: integer
      produces:
      - application/pdf
      responses:
        "200":
          description: Payslip PDF
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Download payslip PDF
      tags:
      - Payroll
  /payrolls/history:
    get:
      description: Return previously generated payroll records for the authenticated
        user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.PayrollResponseDoc'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get payroll history
      tags:
      - Payroll
  /users:
    get:
      description: List employees and managers with their assigned location, shift,
        and base salary.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.UserResponseDoc'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user account and assign role, location, optional shift,
        and base salary.
      parameters:
      - description: User payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateUserRequestDoc'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.UserResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - Users
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"